editoast: authz: define `authz::v2::group` for organization
No code changes.
